refactor(elements): extract FieldResolve+Wait+Discover sub-modules
Phase 12a §7 of the CreateElementMediator god-module split.
Moves three small but tightly cohesive clusters out of
CreateElementMediator.ts into three sibling sub-modules under
Create/.

Why
---
After §6 (Race) the parent still owned three discrete pre-race
concerns: (1) `resolveField`/`resolveClickable` field wrapping
that delegates to resolveFieldPipeline, (2) `waitForLoadingDone`
loading-indicator polling against WK_DASHBOARD.LOADING, and
(3) `discoverForm`/`scopeToForm`/`discoverErrors` form-anchor +
error discovery on top of FormAnchor / FormErrorDiscovery.

None of those clusters share state with the race cluster (they
own pre-race lifecycle concerns) and none of them depend on each
other except through the shared IFormCache type. Bundling them
into one commit clears 223 lines from the parent in a single
landing while keeping the three sub-modules small and focused
(each ≤120 LoC); the alternative — three separate commits —
would each move ~50-80 LoC and front-load the same husky 23-gate
run-time three times instead of once for very modest extra
isolation.

What
----
- Create/FieldResolve.ts (new) — three exports: `IFormCache`,
  `buildResolveField`, `buildResolveClickable`. Five private
  helpers stay encapsulated (IResolveOpts, tryScopedResolve,
  resolveFieldImpl, handleResolveError, resolveFieldForPage).
- Create/Discover.ts (new) — three exports: `buildDiscoverErrors`,
  `buildDiscoverForm`, `buildScopeToForm`. Two private helpers
  stay encapsulated (discoverFormCore, handleDiscoverFormError).
  Re-imports `IFormCache` from FieldResolve.
- Create/Wait.ts (new) — two exports: `buildWaitForLoadingDone`
  and a transparent re-export of `ELEMENTS_LOADING_DELAY_MS` so
  the wait cluster owns its tuning knob (the constant still
  lives in ActionExecutors for the other action callers that
  share the same delay). Two private helpers stay encapsulated
  (isAnyLoadingVisible, waitOnceForLoading).
- Create/index.ts — barrel re-exports the eight new public
  symbols.
- CreateElementMediator.ts — drops the extracted 223-line
  cluster; drops the now-unused imports (`ScraperErrorTypes`
  stays for the navigation-error path, but `none`/`Option`/
  `some`, `WK_DASHBOARD`, `discoverFormAnchor`/`IFormAnchor`/
  `scopeCandidates`, `checkFrameForErrors`/`discoverFormErrors`/
  `IFormErrorScanResult`, `resolveFieldPipeline`,
  `IFieldContext`, `ELEMENTS_LOADING_DELAY_MS`, `isOk` all
  disappear); pulls the eight new symbols from the barrel.
- .github/docs-coverage-allowlist.txt — eight allowlist entries
  (Reason per symbol) for the new internal exports.

Public surface
--------------
lib/index.d.ts + lib/index.d.cts SHA256 BYTE-IDENTICAL to baseline
(both 14823 B, hash 221E1611…AAB055).
lib/index.cjs grew by ~1.3 KB for the three new module init
wrappers — implementation reorganisation, not API change.

Verification
------------
- tsc --noEmit: clean
- eslint Mediator/Elements/Create + parent: clean
- prettier: clean (Wait re-export and barrel re-sort auto-fixed)
- lint:architecture (src/Scrapers/Pipeline): 0 violations
- jest Mediator/Elements (30 suites / 270 tests): all pass
- npm run build: success; d.ts hashes match baseline
